Use helper_test.go
