--- tests/utils_test.go
@@ -1,53 +1,0 @@
-// Copyright 2016-2025, Pulumi Corporation.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package tests
-
-import (
-	"os"
-	"strings"
-	"testing"
-
-	"github.com/pulumi/providertest/pulumitest"
-	"github.com/pulumi/providertest/pulumitest/opttest"
-	"github.com/pulumi/pulumi/pkg/v3/testing/integration"
-)
-
-// Skip the test if it is being run locally without cloud credentials being configured.
-func skipLocalRunsWithoutCreds(t *testing.T) {
-	if _, ci := os.LookupEnv("CI"); ci {
-		return // never skip when in CI
-	}
-
-	awsConfigured := false
-	for _, envVar := range os.Environ() {
-		if strings.HasPrefix(strings.ToUpper(envVar), "AWS_ACCESS_KEY_ID") {
-			awsConfigured = true
-		}
-		if strings.HasPrefix(strings.ToUpper(envVar), "AWS_PROFILE") {
-			awsConfigured = true
-		}
-	}
-	if !awsConfigured {
-		t.Skip("AWS configuration such as AWS_PROFILE env var is required to run this test")
-	}
-}
-
-func newPulumiTest(t pulumitest.PT, source string, opts ...opttest.Option) *pulumitest.PulumiTest {
-	t.Helper()
-	pto := integration.ProgramTestOptions{Dir: source}
-	randomStackName := pto.GetStackName()
-	opts = append(opts, opttest.StackName(string(randomStackName)))
-	return pulumitest.NewPulumiTest(t, source, opts...)
-}
